pyproject.toml
[project]
name = "buildcontrol-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "bcrypt==4.3.0",
    "fabric>=3.2.2",
    "fastapi[standard]>=0.123.8",
    "greenlet>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.3.0",
    "sqlmodel>=0.0.27",
]

[dependency-groups]
dev = [
    "commitizen>=4.10.0",
    "pre-commit>=4.5.0",
]
